refactor: use architecture unsigned integer type instead of i64

+++ src/algorithm/look_and_say.rs
@@ -1,4 +1,4 @@
-type DigitCounter = (char, i64);
+type DigitCounter = (char, usize);
 
 #[allow(dead_code, unused_variables)]
 fn look_and_say(_num_to_look_at: &String) -> String {
